风控/xinyan/requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.1
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
logger==1.4
MarkupSafe==1.1.0
Naked==0.1.31
numpy==1.15.4
pandas==0.23.4
pandasql==0.7.3
pyasn1==0.4.4
pycparser==2.19
pymongo==3.7.2
PyMySQL==0.9.2
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
redis==3.0.1
requests==2.20.1
rsa==4.0
shellescape==3.4.1
six==1.11.0
SQLAlchemy==1.2.14
tornado==5.1.1
urllib3==1.24.1
Werkzeug==0.14.1
